+++ test/sql/viewer.sql
@@ -45,6 +45,7 @@ select emaj.emaj_cleanup_rollback_state();
 select count(*) from emaj.emaj_log_stat_group('myGroup1',NULL,NULL);
 select count(*) from emaj.emaj_log_stat_groups(array['myGroup1'],NULL,NULL);
 select count(*) from emaj.emaj_detailed_log_stat_group('myGroup1',NULL,NULL);
+select count(*) from emaj.emaj_detailed_log_stat_groups(array['myGroup1'],NULL,NULL);
 select emaj.emaj_estimate_rollback_group('myGroup1',emaj.emaj_get_previous_mark_group('myGroup1',current_timestamp),FALSE);
 select emaj.emaj_estimate_rollback_groups(array['myGroup1'],emaj.emaj_get_previous_mark_group('myGroup1',current_timestamp),FALSE);
 select * from emaj.emaj_rollback_activity();